simplify getting decorators so it doesn't need call through multiple static methods
